Rewrite VisualiserLine + add VisualiserContainer.
